+++ src/aichat/aichat.module.ts
@@ -6,6 +6,7 @@ import { PreguntasRepository } from './repositories/preguntas.repository';
 import { ProductsRepository } from '../products/repositories/products.repository';
 import { PrismaModule } from '../prisma/prisma.module';
 import { ConverterService } from './utils/converter.service';
+import { AssistantToolsService } from './utils/assistant-tools.service';
 
 @Module({
   imports: [PrismaModule],
@@ -16,6 +17,7 @@ import { ConverterService } from './utils/converter.service';
     PreguntasRepository,
     ProductsRepository,
     ConverterService,
+    AssistantToolsService,
   ],
   exports: [AichatService],
 })